+++ code/FinanceManager.Domain/Labels/Entities/RecurringCadence.cs
@@ -1,0 +1,9 @@
+namespace FinanceManager.Domain.Labels.Entities;
+
+public enum RecurringCadence
+{
+    Weekly,
+    Monthly,
+    Quarterly,
+    Annual
+}